requirements.txt
django
gunicorn
djangorestframework
psycopg2-binary
django-filter
flake8
